Tiny proof of concept to see how it is to build the docs runner in rust
It shows that we need to parse the markdown ourselves, and serialization
is something we would also have to do separately.
Thus it somehow looks like using bash for this task is not such
a bad idea, even though the most difficult problems seems to
be the markdown rendering, just because I have no clue if
pulldown-cmark can do it.