vault-init uses an explicit --name
This cleans up the CLI a little, as previously we would
mis-use the 'vault-id' of the parent command.
Now it's clear what we are doing thanks to the 'vault init --name'
argument.

This also plays better with partitions, as it's now made clear that
it will primarily affect how partitions are added/removed.

Also it means that interacting with a vault that has partitions should
be easy and not require people to select the leading vault directly.

Let's see how that will work.