Remove CLI parse issues by enforcing using '=' when needed
This is required whenever there is a variable list of positional
arguments, while allowing '--key value' at the same time.
In the latter case, these will only be parsed correctly if
'--key=value' is enforced.
Otherwise one gets 'weird' errors, as it tries to use the option
as positional argument.